LayoutTests/platform/blackberry/accessibility/aria-text-role-expected.txt

hello world this is a test more test
This tests that you can set an ARIA text role and that it will not have children through hit testing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL textrole.role should be AXRole: AXStaticText. Was .
FAIL textrole.stringValue should be AXValue: all at once. Was .
PASS textrole.elementAtPoint(x, y).isEqual(textrole) is true
PASS successfullyParsed is true

TEST COMPLETE
